bnd: Set base in transformer and refactor action into a new package
